requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
Pillow==9.4.0
python-dotenv==1.0.0
python-multipart==0.0.5